+++ android/src/main/java/com/revenuecat/purchases/hybridcommon/OnResult.java
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.common;
+package com.revenuecat.purchases.hybridcommon;
 
 import java.util.Map;
