+++ android/src/main/java/com/revenuecat/purchases/hybridcommon/mappers/OfferingsMapper.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.common.mappers
+package com.revenuecat.purchases.hybridcommon.mappers
 
 import com.revenuecat.purchases.Offering
 import com.revenuecat.purchases.Offerings
