+++ android/src/test/java/com/revenuecat/purchases/hybridcommon/PurchasesPeriodTests.kt
@@ -1,10 +1,7 @@
-package com.revenuecat.purchases.common
+package com.revenuecat.purchases.hybridcommon
 
-import com.revenuecat.purchases.common.mappers.PurchasesPeriod
+import com.revenuecat.purchases.hybridcommon.mappers.PurchasesPeriod
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 
 internal class PurchasesPeriodTests {
